Cleanup defunct factories concept
