+++ Sources/StringGenerator/Snippet.swift
@@ -15,6 +15,14 @@ extension CodeBlockItemListBuilder {
     }
 }
 
+extension CodeBlockItemListBuilder {
+    static func buildExpression(
+        _ expression: some Snippet<some StmtSyntaxProtocol>
+    ) -> CodeBlockItemListBuilder.Component {
+        [CodeBlockItemSyntax(item: .stmt(StmtSyntax(expression.syntax)))]
+    }
+}
+
 extension MemberBlockItemListBuilder {
     static func buildExpression(
         _ expression: some Snippet<some DeclSyntaxProtocol>
